+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.3.17 [unreleased]
+- feat: Support custom transport [PR 84]
 - chore: Implement functions to add session to bitswap [PR 83]
 - refactor: Use channels instead of Subscription [PR 82]
 - feat: Ability to add custom behaviour [PR 81]
@@ -28,6 +29,7 @@
 [PR 81]: https://github.com/dariusc93/rust-ipfs/pull/81
 [PR 82]: https://github.com/dariusc93/rust-ipfs/pull/82
 [PR 83]: https://github.com/dariusc93/rust-ipfs/pull/83
+[PR 84]: https://github.com/dariusc93/rust-ipfs/pull/84
 
 # 0.3.16
 - fix: Return events from gossipsub stream [PR 68]